Restructuring of ReviewForm
A lot of code would have been repeated, therefore form was outsourced to a new component
