+++ frontend/src/App.js
@@ -9,7 +9,7 @@ import LoginPage from "./pages/LoginPage";
 import AuthProvider from "./context/AuthProvider";
 import BookmarkedPage from "./pages/BookmarkedPage";
 import PrivateRoute from "./routing/PrivateRoute";
-import ReviewForm from "./pages/ReviewForm";
+import CreateReviewForm from "./pages/CreateReviewForm";
 import ReviewPage from "./pages/ReviewPage";
 import UpdateReviewForm from "./pages/UpdateReviewForm";
 
@@ -28,7 +28,11 @@ export default function App() {
             <Route component={LoginPage} path={"/login"} />
             <PrivateRoute component={BookmarkedPage} path={"/me/bookmarked"} />
             <PrivateRoute component={ReviewPage} path={"/me/reviews"} />
-            <PrivateRoute component={ReviewForm} path={"/review/:id"} exact />
+            <PrivateRoute
+              component={CreateReviewForm}
+              path={"/review/:id"}
+              exact
+            />
             <PrivateRoute
               component={UpdateReviewForm}
               path={"/review/:id/:reviewId"}
